Added columns for activities and features
Created cells that scrape the unique activities and features using the code that Chris developed. From that code I wrote a nested loop that iterates through the dataframe, and creates rows for each activity and feature. If the feature or activity is present in that particular trail, it adds a 1. If it is not, it leaves the cell as a 0.